Initial rewrite to AuctionSuperApp
